requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
qrcode[pil]==7.4.2
Pillow>=9.0.0
openpyxl>=3.0.10
